requirements.txt
numpy
sounddevice
soundfile
openai-whisper
keyboard
plyer
requests
langdetect
ollama
llm_axe
edge-tts
warnings
pyaudio
tempfile
subprocess
wave